requirements.txt
APScheduler==3.9.1
comtypes==1.1.11
Django==2.1.15
django-pyodbc-azure-2019
django_crispy_forms==1.14.0
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
pdfkit==1.0.0
pyexcel==0.7.0
python_dateutil==2.8.2
tablib==3.2.1
vtracker==1.0.1
xlwt==1.3.0

